requirements.txt
Django==1.6.5
MySQL-python==1.2.3
psycopg2==2.5.3
South==0.8.4
Twisted==13.2.0
Unipath==1.0
Werkzeug==0.8.3
amqp==1.0.12
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
boto==2.4.1
distribute==0.6.45
django-tastypie==0.9.15
django-nose==1.2
django-auth-ldap>=1.2.0
django-ipware==0.0.8
factory-boy==2.3.1
gunicorn
mimeparse==0.1.3
pep8==1.4.4
pika==0.9.13
pydns==2.3.6
python-dateutil==2.1
python-openid==2.2.5
pytz==2013b
requests==1.2.3
simplejson==3.3.0
six==1.3.0
txAMQP==0.6.1
virtualenv==1.6.1
wsgiref==0.1.2
zope.interface==4.1.0
